// src/services/obsidianRestAPI/generated-types.ts
export interface paths {
// ...
  "/commands/": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Get a list of available commands.
     *      */
    get: {
      parameters: {
        query?: never;
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description A list of available commands. */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            /** @example {
             *       "commands": [
             *         {
             *           "id": "global-search:open",
             *           "name": "Search: Search in all files"
             *         },
             *         {
             *           "id": "graph:open",
             *           "name": "Graph view: Open graph view"
             *         }
             *       ]
             *     } */
            "application/json": {
              commands?: {
                id?: string;
                name?: string;
              }[];
            };
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/search/": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /**
     * Search for documents matching a specified search query
     *
     * @description Evaluates a provided query against each file in your vault.
     *
     *     This endpoint supports multiple query formats.  Your query should be specified in your request's body, and will be interpreted according to the `Content-type` header you specify from the below options.Additional query formats may be added in the future.
     *
     *     # Dataview DQL (`application/vnd.olrapi.dataview.dql+txt`)
     *
     *     Accepts a `TABLE`-type Dataview query as a text string.  See [Dataview](https://blacksmithgu.github.io/obsidian-dataview/query/queries/)'s query documentation for information on how to construct a query.
     *
     *     # JsonLogic (`application/vnd.olrapi.jsonlogic+json`)
     *
     *     Accepts a JsonLogic query specified as JSON.  See [JsonLogic](https://jsonlogic.com/operations.html)'s documentation for information about the base set of operators available, but in addition to those operators the following operators are available:
     *
     *     - `glob: [PATTERN, VALUE]`: Returns `true` if a string matches a glob pattern.  E.g.: `{"glob": ["*.foo", "bar.foo"]}` is `true` and `{"glob": ["*.bar", "bar.foo"]}` is `false`.
     *     - `regexp: [PATTERN, VALUE]`: Returns `true` if a string matches a regular expression.  E.g.: `{"regexp": [".*\.foo", "bar.foo"]` is `true` and `{"regexp": [".*\.bar", "bar.foo"]}` is `false`.
     *
     *     Returns only non-falsy results.  "Non-falsy" here treats the following values as "falsy":
     *
     *     - `false`
     *     - `null` or `undefined`
     *     - `0`
     *     - `[]`
     *     - `{}`
     *
     *     Files are represented as an object having the schema described
     *     in the Schema named 'NoteJson' at the bottom of this page.
     *     Understanding the shape of a JSON object from a schema can be
     *     tricky; so you may find it helpful to examine the generated metadata
     *     for individual files in your vault to understand exactly what values
     *     are returned.  To see that, access the `GET` `/vault/{filePath}`
     *     route setting the header:
     *     `Accept: application/vnd.olrapi.note+json`.  See examples below
     *     for working examples of queries performing common search operations.
     *
     */
    post: {
      parameters: {
        query?: never;
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody: {
        content: {
          "application/vnd.olrapi.dataview.dql+txt": string;
          "application/vnd.olrapi.jsonlogic+json": {
            [key: string]: unknown;
          };
        };
      };
      responses: {
        /** @description Success */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            "application/json": {
              /** @description Path to the matching file */
              filename: string;
              result:
                | string
                | number
                | unknown[]
                | Record<string, never>
                | boolean;
            }[];
          };
        };
        /** @description Bad request.  Make sure you have specified an acceptable
         *     Content-Type for your search query.
         *      */
        400: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            "application/json": components["schemas"]["Error"];
          };
        };
      };
    };
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/search/simple/": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** Search for documents matching a specified text query
     *      */
    post: {
      parameters: {
        query: {
          /** @description Your search query */
          query: string;
          /** @description How much context to return around the matching string */
          contextLength?: number;
        };
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Success */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            "application/json": {
              /** @description Path to the matching file */
              filename?: string;
              matches?: {
                context: string;
                match: {
                  end: number;
                  start: number;
                };
              }[];
              score?: number;
            }[];
          };
        };
      };
    };
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/vault/": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * List files that exist in the root of your vault.
     *
     * @description Lists files in the root directory of your vault.
     *
     *     Note: that this is exactly the same API endpoint as the below "List files that exist in the specified directory." and exists here only due to a quirk of this particular interactive tool.
     *
     */
    get: {
      parameters: {
        query?: never;
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Success */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            /** @example {
             *       "files": [
             *         "mydocument.md",
             *         "somedirectory/"
             *       ]
             *     } */
            "application/json": {
              files?: string[];
            };
          };
        };
        /** @description Directory does not exist */
        404: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            "application/json": components["schemas"]["Error"];
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    Error: {
      /**
       * @description A 5-digit error code uniquely identifying this particular type of error.
       *
       * @example 40149
       */
      errorCode?: number;
      /**
       * @description Message describing the error.
       * @example A brief description of the error.
       */
      message?: string;
    };
    NoteJson: {
      content: string;
      frontmatter: {
        [key: string]: unknown;
      };
      path: string;
      stat: {
        ctime: number;
        mtime: number;
        size: number;
      };
      tags: string[];
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}